Add completed marked on project index and :show.
A project is considered completed when it has at least one step, and
all its steps are completed.

As of now, this is checked Just-In-Time each time one project needs
to be checked for completion (so the project index has to load the
steps for all the user's projects for this to be possible).

This will probably lead to performance concerns further down the
road, so better keep this in mind to fix it as soon as it becomes
possible.